refactor: add exclusions to sonarqube
