+++ src/main/java/com/softserveinc/dokazovi/exception/InvalidIdEntityException.java
@@ -1,4 +1,4 @@
-package com.softserveinc.dokazovi.error;
+package com.softserveinc.dokazovi.exception;
 
 public class InvalidIdEntityException extends EntityException {
