+++ src/test/java/com/softserveinc/dokazovi/service/impl/PostServiceImplTest.java
@@ -5,7 +5,7 @@ import com.softserveinc.dokazovi.entity.DirectionEntity;
 import com.softserveinc.dokazovi.entity.PostEntity;
 import com.softserveinc.dokazovi.entity.UserEntity;
 import com.softserveinc.dokazovi.entity.enumerations.PostStatus;
-import com.softserveinc.dokazovi.error.InvalidIdEntityException;
+import com.softserveinc.dokazovi.exception.InvalidIdEntityException;
 import com.softserveinc.dokazovi.mapper.PostMapper;
 import com.softserveinc.dokazovi.repositories.PostRepository;
 import org.junit.jupiter.api.BeforeEach;
